move chainconfig to genesis in utils
